Fix partial retry logic, apply better defaults (#4)
Fluentd has a native retry path for batching plugins (you just have to throw an
exception in your plugin and it'll retry the batch). But previously, we
wouldn't correctly retry if some, but not all, events in a batch were rejected.
Usually this happens because of rate limiting.

Handle this case. Also apply saner configuration defaults with respect to the
batching behavior:
- Limit batch sizes (in bytes) to 500K rather than 8M
- Limit batch delay to 1s rather than 60s
- Limit backoff delay to 30s rather than ~forever

Test plan: new specs