Add presampled_key option for events that have been sampled before getting to fluent
Problem: we have some applications which are doing sampling internally,
and want to forward those records along to honeycomb, with the correct
sampling rate.

This addition allows users to set an optional `presampled_key` in their
configuration, which will be used as the `samplerate` sent to Honeycomb.
If an event doesn't include the `presampled_key`, the existing logic is
still used (all events sent with `samplerate` 1, or N events are sent
with whatever `sample_rate` is set to).