Ignore bad presampled_key sample rates
As discussed in pull request #9, if using `presampled_key`, an event may
be emitted with a bad value. This adds basic checking on the that value.
Specifically, any non-positive integer is ignored.

Should an event have its `presampled_key` ignored, the event will be
passed through to Honeycomb, with the `samplerate` of 1, regardless of
any configuration set `sample_rate`, because the intended action would
appear to be sending the event in full to Honeycomb, even if the weight
is wrong.