Retry rather than raise exception when batch fails
When a non-200 error is returned by our batch endpoint, we raise an exception. I don't know what the original reason was for raising an exception here, but currently it causes the plugin to exit abnormally. Instead, we should return the entire batch, which will result in a retry of all events in the batch.